refactor(proof): nest SP1 host helpers under zk-backend host module
Keep the Succinct ZkProver backends at the crate surface and group leftover
stdin, ELF, cluster, and network helpers under succinct/host.
