+++ crates/proof/zk/backend/src/succinct/host/contract.rs
