+++ crates/proof/zk/backend/src/succinct/host/stats.rs
