+++ crates/proof/zk/backend/src/succinct/provider.rs
@@ -2,7 +2,7 @@
 
 use std::{error::Error as StdError, fmt, sync::Arc};
 
-use crate::succinct::{get_agg_proof_stdin, get_sp1_stdin};
+use super::host::{get_agg_proof_stdin, get_sp1_stdin};
 use alloy_primitives::{Address, B256};
 use base_l1_head::{L1HeadCalculator, L1HeadError};
 use base_proof_host::Metrics;
